broke add record page to components
